pyproject.toml
[project]
name = "personal_financial_accounting"
version = "0.1.0"
description = ""
authors = [
    {name = "Yauheni Mukhin",email = "evgen3email@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pip (>=25.1.1,<26.0.0)",
    "django-rest-swagger (>=2.2.0,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
